requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0

# Web server
flask>=2.0.0
flask-cors>=3.0.0

# Graph visualization
networkx>=3.0
pyvis>=0.3.0

# Utilities
termcolor>=2.0.0
python-dotenv>=1.0.0

# Optional: LangSmith for tracing
langsmith>=0.3.0

